Crate socket

Structs

Socket

Constants

AF_INET

AF_INET6

IPPROTO_IP

IPPROTO_IPV6

IPPROTO_RAW

IPPROTO_TCP

IPV6_ADD_MEMBERSHIP

IPV6_DROP_MEMBERSHIP

IP_ADD_MEMBERSHIP

IP_DROP_MEMBERSHIP

IP_HDRINCL

IP_MULTICAST_LOOP

IP_MULTICAST_TTL

IP_TTL

SHUT_RD

SHUT_WR

SOCK_DGRAM

SOCK_RAW

SOCK_STREAM

SOL_SOCKET

SO_BROADCAST

SO_ERROR

SO_KEEPALIVE

SO_REUSEADDR

TCP_NODELAY

Functions

htonl
Convert a value from host byte order to network byte order

htons
Convert a value from host byte order to network byte order

ntohl
Convert a value from network byte order to host byte order

ntohs
Convert a value from network byte order to host byte order